+++ AddOns/Cell/CHANGELOG.md
@@ -1,4 +1,3 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r178-release...d7a8b1877feddb565b97d3d146c51dad5e0805ed)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r180-release...4a2f0ed2fe8b77f43f4428138cf8f5b81187f85b)
 
-- Implement translit
-- Update PW:S indicator
+- Add a "Shape" option for PW:S